initial implementation sensor info view
- replaces the old image-based sensor countdown (only for Libre)
- will now work for all CGM systems that report a sensor start date (or sensor age in the case of Libre)
- if no current/valid data is found, the view remains hidden
- currently showing for Libre 2 EU BLE (14.5 days), compatible Libre via external transmitter (14.5 days), Dexcom G5/G6/One (10.0 days) and LibreLinkUp follower (14.0 days)
- starts at 0% progress for 0d and finishes at 100% progress for maxAge as per above line
- To simply UI, sensor countdown options are removed completely
- All maxAge values are in Constants file for easy updating if ever needed

+++ xdrip/BluetoothTransmitter/CGM/Libre/MiaoMiao/CGMMiaoMiaoTransmitter.swift
@@ -327,7 +327,7 @@ class CGMMiaoMiaoTransmitter:BluetoothTransmitter, CGMTransmitter {
         return .miaomiao
     }
     
-    func maxSensorAgeInDays() -> Int? {
+    func maxSensorAgeInDays() -> Double? {
         
         return libreSensorType?.maxSensorAgeInDays()
